+++ core/src/com/geometric/wars/utils/Direction2D.java
@@ -1,4 +1,4 @@
-package com.geometric.wars;
+package com.geometric.wars.utils;
 
 import com.badlogic.gdx.math.Vector2;
