+++ core/src/com/geometric/wars/utils/DirectionConversionException.java
@@ -1,4 +1,4 @@
-package com.geometric.wars;
+package com.geometric.wars.utils;
 
 public class DirectionConversionException extends RuntimeException {
 }
