+++ core/src/com/geometric/wars/utils/TypeOfGame.java
@@ -1,4 +1,4 @@
-package com.geometric.wars;
+package com.geometric.wars.utils;
 
 public enum TypeOfGame {
     MOCK_GAME,
