+++ core/src/com/geometric/wars/utils/axes/CoordinateAxes3D.java
@@ -1,4 +1,4 @@
-package com.geometric.wars.debug.axes;
+package com.geometric.wars.utils.axes;
 
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.g3d.Environment;
@@ -6,11 +6,11 @@ import com.badlogic.gdx.graphics.g3d.ModelBatch;
 import com.badlogic.gdx.math.Vector3;
 
 public class CoordinateAxes3D {
-    private com.geometric.wars.debug.axes.CoordinateAxis axisX, axisY, axisZ;
+    private CoordinateAxis axisX, axisY, axisZ;
     public CoordinateAxes3D() {
-        axisX = new com.geometric.wars.debug.axes.CoordinateAxis();
-        axisY = new com.geometric.wars.debug.axes.CoordinateAxis();
-        axisZ = new com.geometric.wars.debug.axes.CoordinateAxis();
+        axisX = new CoordinateAxis();
+        axisY = new CoordinateAxis();
+        axisZ = new CoordinateAxis();
 
         axisX.setColor(Color.RED);
         axisY.setColor(Color.GREEN);
